+++ src/coordinates.h
@@ -19,28 +19,28 @@
 // Coordinate parameters.  See 
 extern int use_eKS_internal;
 extern int metric;
-extern double a, hslope; // mks
-extern double poly_norm, poly_xt, poly_alpha, mks_smooth; // fmks
-extern double mks3R0, mks3H0, mks3MY1, mks3MY2, mks3MP0; // mks3
-extern double startx[NDIM], stopx[NDIM], dx[NDIM];
-extern double R0, Rin, Rout, Rh;
+extern REAL a, hslope; // mks
+extern REAL poly_norm, poly_xt, poly_alpha, mks_smooth; // fmks
+extern REAL mks3R0, mks3H0, mks3MY1, mks3MY2, mks3MP0; // mks3
+extern REAL startx[NDIM], stopx[NDIM], dx[NDIM];
+extern REAL R0, Rin, Rout, Rh;
 
-void bl_coord(double *X, double *r, double *th);
-void bl_to_ks(double X[NDIM], double ucon_bl[NDIM], double ucon_ks[NDIM]);
-void ks_to_bl(double X[NDIM], double ucon_ks[NDIM], double ucon_bl[NDIM]);
-void gcov_func(double *X, double gcov[][NDIM]);
+void bl_coord(REAL *X, REAL *r, REAL *th);
+void bl_to_ks(REAL X[NDIM], REAL ucon_bl[NDIM], REAL ucon_ks[NDIM]);
+void ks_to_bl(REAL X[NDIM], REAL ucon_ks[NDIM], REAL ucon_bl[NDIM]);
+void gcov_func(REAL *X, REAL gcov[][NDIM]);
 // TODO privatize these, why are they needed in models?
-void gcov_ks(double r, double th, double gcov[NDIM][NDIM]);
-void gcov_bl(double r, double th, double gcov[NDIM][NDIM]);
+void gcov_ks(REAL r, REAL th, REAL gcov[NDIM][NDIM]);
+void gcov_bl(REAL r, REAL th, REAL gcov[NDIM][NDIM]);
 
 // Internal
-void set_dxdX(double X[NDIM], double dxdX[NDIM][NDIM]);
-void set_dXdx(double X[NDIM], double dXdx[NDIM][NDIM]);
-void vec_to_ks(double X[NDIM], double v_nat[NDIM], double v_ks[NDIM]);
-void vec_from_ks(double X[NDIM], double v_ks[NDIM], double v_nat[NDIM]);
+void set_dxdX(REAL X[NDIM], REAL dxdX[NDIM][NDIM]);
+void set_dXdx(REAL X[NDIM], REAL dXdx[NDIM][NDIM]);
+void vec_to_ks(REAL X[NDIM], REAL v_nat[NDIM], REAL v_ks[NDIM]);
+void vec_from_ks(REAL X[NDIM], REAL v_ks[NDIM], REAL v_nat[NDIM]);
 
 // Translation to native coords
-void native_coord(double r, double th, double phi, double X[NDIM]);
-double root_find(double X[NDIM]);
+void native_coord(REAL r, REAL th, REAL phi, REAL X[NDIM]);
+REAL root_find(REAL X[NDIM]);
 
 #endif // COORDINATES_H
