+++ src/model_tetrads.h
@@ -10,12 +10,12 @@
 
 #include "decs.h"
 
-int make_camera_tetrad(double X[NDIM], double Econ[NDIM][NDIM],
-                    double Ecov[NDIM][NDIM]);
-int make_camera_tetrad_old(double X[NDIM], double Econ[NDIM][NDIM],
-                        double Ecov[NDIM][NDIM]);
-int make_plasma_tetrad(double Ucon[NDIM], double Kcon[NDIM], double Bcon[NDIM],
-                    double Gcov[NDIM][NDIM], double Econ[NDIM][NDIM],
-                    double Ecov[NDIM][NDIM]);
+int make_camera_tetrad(REAL X[NDIM], REAL Econ[NDIM][NDIM],
+                    REAL Ecov[NDIM][NDIM]);
+int make_camera_tetrad_old(REAL X[NDIM], REAL Econ[NDIM][NDIM],
+                        REAL Ecov[NDIM][NDIM]);
+int make_plasma_tetrad(REAL Ucon[NDIM], REAL Kcon[NDIM], REAL Bcon[NDIM],
+                    REAL Gcov[NDIM][NDIM], REAL Econ[NDIM][NDIM],
+                    REAL Ecov[NDIM][NDIM]);
 
 #endif /* MODEL_TETRADS_H */
