+++ src/radiation.h
@@ -11,10 +11,10 @@
 #include "decs.h"
 
 /* radiation */
-double Bnu_inv(double nu, double Thetae);
-double jnu_inv(double nu, double Thetae, double Ne, double B, double theta);
-double get_fluid_nu(double Kcon[NDIM], double Ucov[NDIM]);
-double get_bk_angle(double X[NDIM], double Kcon[NDIM], double Ucov[NDIM],
-              double Bcon[NDIM], double Bcov[NDIM]);
+REAL Bnu_inv(REAL nu, REAL Thetae);
+REAL jnu_inv(REAL nu, REAL Thetae, REAL Ne, REAL B, REAL theta);
+REAL get_fluid_nu(REAL Kcon[NDIM], REAL Ucov[NDIM]);
+REAL get_bk_angle(REAL X[NDIM], REAL Kcon[NDIM], REAL Ucov[NDIM],
+              REAL Bcon[NDIM], REAL Bcov[NDIM]);
 
 #endif /* RADIATION_H */
